fix commit

+++ app/controllers/LoadDataController.php
@@ -4,7 +4,48 @@ class LoadDataController extends \Controller {
 	
 	public function index()
 	{
-        $csvFile = public_path().'/tablas/usuarios.csv';
+        DB::table('compras')->delete();
+        // $this->usuarios();
+         // $this->categorias();
+         // $this->marcas();
+
+         //     $precio_venta = new PrecioVenta;
+         //     $precio_venta->id = 1;
+         //     $precio_venta->nombre = 'Base';
+         //     $precio_venta->save();
+
+        //$this->productos(1);
+       // $this->productos(2);
+
+        // $this->clientes(1);
+        // $this->clientes(2);
+        // $this->clientes(3);
+
+        // $this->soporte();
+        // $this->gastos();
+        // $this->egresos();
+        // $this->proveedores();
+
+    }
+
+
+    public function readCSV($csvFile)
+    {
+        $file_handle = fopen($csvFile, 'r');
+
+        while (!feof($file_handle))
+        {
+            $line_of_text[] = fgetcsv($file_handle, 1024);
+        }
+
+        fclose($file_handle);
+
+        return $line_of_text;
+    }
+
+    public function usuarios()
+    {
+        $csvFile = app_path().'/tablas/usuarios.csv';
 
         $items = $this->readCSV($csvFile);
 
@@ -12,7 +53,7 @@ class LoadDataController extends \Controller {
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $user = new User;
             $user->id     = $item[0];
@@ -29,7 +70,11 @@ class LoadDataController extends \Controller {
         echo $counter. " " ."usuarios ingresados!";
         echo "<br>";
 
-       $csvFile = public_path().'/tablas/categorias.csv';
+    }
+
+    public function categorias()
+    {
+        $csvFile = app_path().'/tablas/categorias.csv';
 
         $items = $this->readCSV($csvFile);
 
@@ -39,7 +84,7 @@ class LoadDataController extends \Controller {
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $categoria = new Categoria;
             $categoria->id     = $item[0];
@@ -47,10 +92,14 @@ class LoadDataController extends \Controller {
             $categoria->save();
 
         }
-        echo $counter. " " ."marcas ingresadas!";
+        echo $counter. " " ."categoria ingresadas!";
         echo "<br>";
 
-       $csvFile = public_path().'/tablas/marca.csv';
+    }
+
+    public function marcas()
+    {
+        $csvFile = app_path().'/tablas/marca.csv';
 
         $items = $this->readCSV($csvFile);
 
@@ -60,7 +109,7 @@ class LoadDataController extends \Controller {
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $marca = new Marca;
             $marca->id     = $item[0];
@@ -70,24 +119,27 @@ class LoadDataController extends \Controller {
         }
         echo $counter. " " ."marcas ingresadas!";
         echo "<br>";
+    }
 
-
-            $precio_venta = new PrecioVenta;
-            $precio_venta->descripcion = 'prueba';
-            $precio_venta->save();
-
-
-        $csvFile = public_path().'/tablas/producto.csv';
-
+    public function productos($num)
+    {
+        if ($num == 1) 
+        {
+            $csvFile = app_path().'/tablas/producto.csv';
+            DB::table('productos')->delete();
+        }
+        if($num == 2)
+        {
+            $csvFile = app_path().'/tablas/producto1.csv';
+        }
+        
         $items = $this->readCSV($csvFile);
 
         $counter = 0;
 
-        DB::table('productos')->delete();
-
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $producto = new Producto;
             $producto->id     = $item[0];
@@ -103,18 +155,33 @@ class LoadDataController extends \Controller {
 
         echo $counter. " " ."productos ingresados!";
         echo "<br>";
+    }
 
-       $csvFile = public_path().'/tablas/cliente.csv';
+    public function clientes($num)
+    {
+        if ($num == 1) 
+        {
+             $csvFile = app_path().'/tablas/cliente.csv';
+             DB::table('clientes')->delete();
+        }
+        if ($num == 2) 
+        {
+             $csvFile = app_path().'/tablas/cliente1.csv';
+        }
+        if ($num == 3) 
+        {
+             $csvFile = app_path().'/tablas/cliente2.csv';
+        }
 
         $items = $this->readCSV($csvFile);
 
         $counter = 0;
 
-        DB::table('clientes')->delete();
+       
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $cliente = new Cliente;
             $cliente->id     = $item[0];
@@ -129,8 +196,11 @@ class LoadDataController extends \Controller {
 
         echo $counter. " " ."clientes ingresados!";
         echo "<br>";
+    }
 
-       $csvFile = public_path().'/tablas/soporte.csv';
+    public function soporte()
+    {
+        $csvFile = app_path().'/tablas/soporte.csv';
 
         $items = $this->readCSV($csvFile);
 
@@ -140,7 +210,7 @@ class LoadDataController extends \Controller {
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $soporte = new Soporte;
             $soporte->id     = $item[0];
@@ -154,7 +224,7 @@ class LoadDataController extends \Controller {
         echo $counter. " " ."soporte ingresados!";
         echo "<br>";
 
-       $csvFile = public_path().'/tablas/detallesoporte.csv';
+       $csvFile = app_path().'/tablas/detallesoporte.csv';
 
         $items = $this->readCSV($csvFile);
 
@@ -164,7 +234,7 @@ class LoadDataController extends \Controller {
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $detalle_soporte = new DetalleSoporte;
             $detalle_soporte->id     = $item[0];
@@ -177,8 +247,114 @@ class LoadDataController extends \Controller {
 
         echo $counter. " " ."detalle_soporte ingresados!";
         echo "<br>";
+    }
 
-       $csvFile = public_path().'/tablas/proveedor.csv';
+    public function gastos()
+    {
+         $csvFile = app_path().'/tablas/gastos.csv';
+
+        $items = $this->readCSV($csvFile);
+
+        $counter = 0;
+
+        DB::table('gastos')->delete();
+
+        foreach ($items as $item) {
+
+            $counter++;
+
+            $gastos = new Gasto;
+            $gastos->id     = $item[0];
+            $gastos->user_id = $item[2];
+            $gastos->tienda_id = $item[1];
+            $gastos->created_at = $item[3];
+            $gastos->save();
+
+        }
+
+        echo $counter. " " ."Gastos ingresados!";
+        echo "<br>";
+
+       $csvFile = app_path().'/tablas/detallegastos.csv';
+
+        $items = $this->readCSV($csvFile);
+
+        $counter = 0;
+
+        DB::table('detalle_gastos')->delete();
+
+        foreach ($items as $item) {
+
+            $counter++;
+
+            $detalle_gasto = new DetalleGasto;
+            $detalle_gasto->id     = $item[0];
+            $detalle_gasto->descripcion = $item[2];
+            $detalle_gasto->monto = $item[3];
+            $detalle_gasto->gasto_id = $item[1];
+            $detalle_gasto->metodo_pago_id = 1;
+            $detalle_gasto->save();
+        }
+
+        echo $counter. " " ."detalle_gastos ingresados!";
+        echo "<br>";
+    }
+
+    public function egresos()
+    { 
+         $csvFile = app_path().'/tablas/egresos.csv';
+
+        $items = $this->readCSV($csvFile);
+
+        $counter = 0;
+
+        DB::table('egresos')->delete();
+
+        foreach ($items as $item) {
+
+            $counter++;
+
+            $egreso = new Egreso;
+            $egreso->id     = $item[0];
+            $egreso->user_id = $item[2];
+            $egreso->tienda_id = $item[1];
+            $egreso->created_at = $item[3];
+            $egreso->save();
+
+        }
+
+        echo $counter. " " ."egresos ingresados!";
+        echo "<br>";
+
+       $csvFile = app_path().'/tablas/egresosdetalle.csv';
+
+        $items = $this->readCSV($csvFile);
+
+        $counter = 0;
+
+        DB::table('detalle_egresos')->delete();
+
+        foreach ($items as $item) {
+
+            $counter++;
+
+            $detalle_egreso = new DetalleEgreso;
+            $detalle_egreso->id     = $item[0];
+            $detalle_egreso->descripcion = $item[2];
+            $detalle_egreso->monto = $item[3];
+            $detalle_egreso->egreso_id = $item[1];
+            $detalle_egreso->metodo_pago_id = 1;
+            $detalle_egreso->save();
+        }
+
+        echo $counter. " " ."detalle_egresos ingresados!";
+        echo "<br>";
+    }
+
+    public function proveedores()
+    {
+
+       $csvFile = app_path().'/tablas/proveedor.csv';
 
         $items = $this->readCSV($csvFile);
 
@@ -188,7 +364,7 @@ class LoadDataController extends \Controller {
 
         foreach ($items as $item) {
 
-        	$counter++;
+            $counter++;
 
             $proveedor = new Proveedor;
             $proveedor->id     = $item[0];
@@ -201,20 +377,5 @@ class LoadDataController extends \Controller {
 
         echo $counter. " " ."proveedores ingresados!";
         echo "<br>";
-    }
-
-
-    public function readCSV($csvFile)
-    {
-        $file_handle = fopen($csvFile, 'r');
-
-        while (!feof($file_handle))
-        {
-            $line_of_text[] = fgetcsv($file_handle, 1024);
-        }
-
-        fclose($file_handle);
-
-        return $line_of_text;
     }
 }
